hooks: remove /etc/apt/sources.list.d/proposed.list
The "bionic-base.tar.xz" we get from cdimage is build with
bionic-prposed enabled by default. This is problematic for us
because core18 is meant to be build from stable only.

This PR removes it for all the extra pacakges we install. We
need to double check with foundations that the bionic-base.tar.xz
is also build from stable only but that is outside the scope of
this PR.