Add motd and issue to writable /etc files
As an embedded developer, I'd like to be able to modify my `motd` and `issue` banners with a config hook when installing my Ubuntu Core image. Because these files are unwritable currently, It appears I straight-up cannot do that, so, I'd like to add them to the `writable` directory so that they can be modified.